+++ mopack/builders/__init__.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+
 from pkg_resources import load_entry_point
 
 from ..base_options import BaseOptions, OptionsHolder
@@ -33,6 +34,37 @@ class Builder(OptionsHolder):
         return result
 
     def path_bases(self):
+        return ()
+
+    def path_values(self, metadata):
+        return {}
+
+    def filter_linkage(self, linkage):
+        return linkage
+
+    def clean(self, metadata, pkg):
+        pass
+
+    def build(self, metadata, pkg):
+        pass
+
+    def deploy(self, metadata, pkg):
+        pass
+
+    def __repr__(self):
+        return '<{}({!r})>'.format(type(self).__name__, self.name)
+
+
+@GenericFreezeDried.fields(rehydrate={'child_builder': Builder})
+class ConfiguringBuilder(Builder):
+    def __init__(self, pkg, *, build=True, _symbols, _child_builder, **kwargs):
+        super().__init__(pkg, **kwargs)
+        if build:
+            self.child_builder = _child_builder(pkg, _symbols=_symbols)
+        else:
+            self.child_builder = None
+
+    def path_bases(self):
         return ('builddir',)
 
     def path_values(self, metadata):
@@ -40,15 +72,17 @@ class Builder(OptionsHolder):
                                                 self.name))
         return {'builddir': builddir}
 
-    def filter_linkage(self, linkage):
-        return linkage
-
     def clean(self, metadata, pkg):
         path_values = pkg.path_values(metadata)
         shutil.rmtree(path_values['builddir'], ignore_errors=True)
 
-    def __repr__(self):
-        return '<{}({!r})>'.format(type(self).__name__, self.name)
+    def build(self, metadata, pkg):
+        if self.child_builder:
+            self.child_builder.build(metadata, pkg)
+
+    def deploy(self, metadata, pkg):
+        if self.child_builder:
+            self.child_builder.deploy(metadata, pkg)
 
 
 class BuilderOptions(GenericFreezeDried, BaseOptions):